ДЗ: 7.10. Отрисуй меня полностью (часть 2): Восстановил два модуля, но теперь один цикл отрисовки
