+++ js/main.js
@@ -1,6 +1,6 @@
 //import './setup.js'; // это файл для настроек загрузки фото, фильтации, редактирования и т.д.
-//import './miniature.js'; // это файл для  отрисовки миниатюр
-//import './big-picture.js'; // это файл для отрисовки окна с полноразмерным изображением
-import './picture.js'; // это файл для отрисовки окна с полноразмерным изображением
+import './miniature.js'; // это файл для  отрисовки миниатюр
+import './big-picture.js'; // это файл для отрисовки окна с полноразмерным изображением
+//import './picture.js'; // это файл для отрисовки окна с полноразмерным изображением
 //import {createPhoto} from './data.js';
